requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
requests==2.21.0
numpy==1.16.1
google-cloud-bigquery==1.10.0
google-cloud-storage==1.19.0
